+++ hypervisor/arch/x86/cpu_caps.c
@@ -36,10 +36,6 @@ struct cpu_capability {
 
 struct cpuinfo_x86 boot_cpu_data;
 
-static bool skip_l1dfl_vmentry;
-static uint64_t x86_arch_capabilities;
-int32_t ibrs_type;
-
 bool cpu_has_cap(uint32_t bit)
 {
 	uint32_t feat_idx = bit >> 5U;
@@ -252,30 +248,6 @@ void init_cpu_capabilities(void)
 	}
 
 	detect_cpu_cap();
-
-	/* For speculation defence.
-	 * The default way is to set IBRS at vmexit and then do IBPB at vcpu
-	 * context switch(ibrs_type == IBRS_RAW).
-	 * Now provide an optimized way (ibrs_type == IBRS_OPT) which set
-	 * STIBP and do IBPB at vmexit,since having STIBP always set has less
-	 * impact than having IBRS always set. Also since IBPB is already done
-	 * at vmexit, it is no necessary to do so at vcpu context switch then.
-	 */
-	ibrs_type = IBRS_NONE;
-
-	/* Currently for APL, if we enabled retpoline, then IBRS should not
-	 * take effect
-	 * TODO: add IA32_ARCH_CAPABILITIES[1] check, if this bit is set, IBRS
-	 * should be set all the time instead of relying on retpoline
-	 */
-#ifndef CONFIG_RETPOLINE
-	if (cpu_has_cap(X86_FEATURE_IBRS_IBPB)) {
-		ibrs_type = IBRS_RAW;
-		if (cpu_has_cap(X86_FEATURE_STIBP)) {
-			ibrs_type = IBRS_OPT;
-		}
-	}
-#endif
 }
 
 static bool is_ept_supported(void)
@@ -332,43 +304,6 @@ void init_cpu_model_name(void)
 		(uint32_t *)(&boot_cpu_data.model_name[44]));
 
 	boot_cpu_data.model_name[48] = '\0';
-}
-
-bool check_cpu_security_cap(void)
-{
-	if (cpu_has_cap(X86_FEATURE_ARCH_CAP)) {
-		x86_arch_capabilities = msr_read(MSR_IA32_ARCH_CAPABILITIES);
-		skip_l1dfl_vmentry = ((x86_arch_capabilities
-			& IA32_ARCH_CAP_SKIP_L1DFL_VMENTRY) != 0UL);
-	} else {
-		return false;
-	}
-
-	if ((!cpu_has_cap(X86_FEATURE_L1D_FLUSH)) && (!skip_l1dfl_vmentry)) {
-		return false;
-	}
-
-	if (!cpu_has_cap(X86_FEATURE_IBRS_IBPB) &&
-		!cpu_has_cap(X86_FEATURE_STIBP)) {
-		return false;
-	}
-
-	return true;
-}
-
-void cpu_l1d_flush(void)
-{
-	/*
-	 * 'skip_l1dfl_vmentry' will be true on platform that
-	 * is not affected by L1TF.
-	 *
-	 */
-	if (!skip_l1dfl_vmentry) {
-		if (cpu_has_cap(X86_FEATURE_L1D_FLUSH)) {
-			msr_write(MSR_IA32_FLUSH_CMD, IA32_L1D_FLUSH);
-		}
-	}
-
 }
 
 static inline bool is_vmx_disabled(void)
