+++ hypervisor/include/arch/x86/cpu.h
@@ -246,15 +246,6 @@ enum pcpu_boot_state {
 	PCPU_STATE_DEAD,
 };
 
-#ifdef STACK_PROTECTOR
-struct stack_canary {
-	/* Gcc generates extra code, using [fs:40] to access canary */
-	uint8_t reserved[40];
-	uint64_t canary;
-};
-void __stack_chk_fail(void);
-#endif
-
 /* Function prototypes */
 void cpu_do_idle(void);
 void cpu_dead(void);
